back_end/requirements.txt
fastapi
uvicorn
qdrant-client
playwright
beautifulsoup4
nltk
python-dotenv
webdriver-manager
pandas
google-generativeai==0.7.1
google-api-core==2.19.0
proto-plus==1.23.0
protobuf==4.24.4
langchain==0.2.6
langchain-google-genai
langchain-qdrant==0.1.1


